Refactor modules, add CLI tests, amd update tooling
