+++ src/extract/isc.rs
@@ -1,29 +1,10 @@
 use anyhow::Result;
-use std::collections::HashSet;
 use xmltree::Element;
 
-use crate::xml_helpers::{find_descendant_ci, get_child_ci};
+use crate::xml_helpers::get_child_ci;
 use crate::{
     IscDhcpOptionsV4, IscDhcpOptionsV6, IscRangeV4, IscRangeV6, IscStaticMap, IscStaticMapV6,
-    Subnet, SubnetV6,
 };
-use std::collections::HashMap;
-use std::net::{Ipv4Addr, Ipv6Addr};
-use std::str::FromStr;
-
-/// Check if Kea DHCPv4 is configured (recursive search)
-pub(crate) fn has_kea_dhcp4(root: &Element) -> bool {
-    find_descendant_ci(root, "Kea")
-        .and_then(|kea| find_descendant_ci(kea, "dhcp4"))
-        .is_some()
-}
-
-/// Check if Kea DHCPv6 is configured (recursive search)
-pub(crate) fn has_kea_dhcp6(root: &Element) -> bool {
-    find_descendant_ci(root, "Kea")
-        .and_then(|kea| find_descendant_ci(kea, "dhcp6"))
-        .is_some()
-}
 
 /// Extract ISC static mappings from the XML tree
 pub fn extract_isc_mappings(root: &Element) -> Result<Vec<IscStaticMap>> {
@@ -333,242 +314,4 @@ pub fn extract_isc_ranges_v6(root: &Element) -> Result<Vec<IscRangeV6>> {
     }
 
     Ok(ranges)
-}
-
-/// Extract interface IPv4 CIDRs from the XML tree (interface name -> CIDR)
-pub fn extract_interface_cidrs(root: &Element) -> Result<HashMap<String, String>> {
-    let mut cidrs = HashMap::new();
-
-    if let Some(interfaces) = get_child_ci(root, "interfaces") {
-        for iface_node in interfaces.children.iter() {
-            if let Some(iface_elem) = iface_node.as_element() {
-                let iface_name = iface_elem.name.clone();
-                let ipaddr = get_child_ci(iface_elem, "ipaddr")
-                    .and_then(|e| e.get_text())
-                    .map(|s| s.to_string())
-                    .unwrap_or_default();
-                let subnet = get_child_ci(iface_elem, "subnet")
-                    .and_then(|e| e.get_text())
-                    .map(|s| s.to_string())
-                    .unwrap_or_default();
-
-                if ipaddr.is_empty() || subnet.is_empty() {
-                    continue;
-                }
-                if ipaddr.eq_ignore_ascii_case("dhcp") {
-                    continue;
-                }
-
-                let prefix: u8 = match subnet.parse() {
-                    Ok(v) => v,
-                    Err(_) => continue,
-                };
-                let ip = match Ipv4Addr::from_str(&ipaddr) {
-                    Ok(ip) => ip,
-                    Err(_) => continue,
-                };
-
-                let net = ipnet::Ipv4Net::new(ip, prefix)
-                    .map_err(|_| crate::MigrationError::InvalidCidr(subnet.clone()))?;
-                let cidr = format!("{}/{}", net.network(), net.prefix_len());
-                cidrs.insert(iface_name, cidr);
-            }
-        }
-    }
-
-    Ok(cidrs)
-}
-
-/// Extract interface IPv6 CIDRs from the XML tree (interface name -> CIDR)
-pub fn extract_interface_cidrs_v6(root: &Element) -> Result<HashMap<String, String>> {
-    let mut cidrs = HashMap::new();
-
-    if let Some(interfaces) = get_child_ci(root, "interfaces") {
-        for iface_node in interfaces.children.iter() {
-            if let Some(iface_elem) = iface_node.as_element() {
-                let iface_name = iface_elem.name.clone();
-                let ipaddr = get_child_ci(iface_elem, "ipaddrv6")
-                    .and_then(|e| e.get_text())
-                    .map(|s| s.to_string())
-                    .unwrap_or_default();
-                let subnet = get_child_ci(iface_elem, "subnetv6")
-                    .and_then(|e| e.get_text())
-                    .map(|s| s.to_string())
-                    .unwrap_or_default();
-
-                if ipaddr.is_empty() || subnet.is_empty() {
-                    continue;
-                }
-                let ipaddr_lower = ipaddr.to_ascii_lowercase();
-                if ipaddr_lower == "dhcp6" || ipaddr_lower == "track6" {
-                    continue;
-                }
-
-                let prefix: u8 = match subnet.parse() {
-                    Ok(v) => v,
-                    Err(_) => continue,
-                };
-                let ip = match Ipv6Addr::from_str(&ipaddr) {
-                    Ok(ip) => ip,
-                    Err(_) => continue,
-                };
-
-                let net = ipnet::Ipv6Net::new(ip, prefix)
-                    .map_err(|_| crate::MigrationError::InvalidCidr(subnet.clone()))?;
-                let cidr = format!("{}/{}", net.network(), net.prefix_len());
-                cidrs.insert(iface_name, cidr);
-            }
-        }
-    }
-
-    Ok(cidrs)
-}
-
-/// Extract Kea subnet4 entries from the XML tree
-/// Supports two schema variations:
-/// 1. <Kea><dhcp4><subnets><subnet4>... (standard OPNsense)
-/// 2. <Kea><dhcp4><subnet4>... (fallback for plugin variations)
-pub fn extract_kea_subnets(root: &Element) -> Result<Vec<Subnet>> {
-    let mut subnets = Vec::new();
-
-    // Navigate to <Kea>/<kea> (case-insensitive) -> <dhcp4>
-    if let Some(kea) = find_descendant_ci(root, "Kea") {
-        if let Some(dhcp4) = find_descendant_ci(kea, "dhcp4") {
-            // Try standard path: <dhcp4><subnets><subnet4>
-            let subnet_container = if let Some(subnets_node) = get_child_ci(dhcp4, "subnets") {
-                Some(subnets_node)
-            } else {
-                // Fallback: <subnet4> directly under <dhcp4>
-                Some(dhcp4)
-            };
-
-            if let Some(container) = subnet_container {
-                for child in container.children.iter() {
-                    if let Some(subnet4) = child.as_element() {
-                        if subnet4.name.eq_ignore_ascii_case("subnet4") {
-                            if let Some(uuid) = subnet4.attributes.get("uuid") {
-                                if let Some(subnet_elem) = get_child_ci(subnet4, "subnet") {
-                                    if let Some(cidr) = subnet_elem.get_text() {
-                                        subnets.push(Subnet {
-                                            uuid: uuid.to_string(),
-                                            cidr: cidr.to_string(),
-                                        });
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(subnets)
-}
-
-/// Extract Kea subnet6 entries from the XML tree
-/// Supports <Kea><dhcp6><subnets><subnet6>
-pub fn extract_kea_subnets_v6(root: &Element) -> Result<Vec<SubnetV6>> {
-    let mut subnets = Vec::new();
-
-    if let Some(kea) = find_descendant_ci(root, "Kea") {
-        if let Some(dhcp6) = find_descendant_ci(kea, "dhcp6") {
-            if let Some(subnets_node) = get_child_ci(dhcp6, "subnets") {
-                for child in subnets_node.children.iter() {
-                    if let Some(subnet6) = child.as_element() {
-                        if subnet6.name.eq_ignore_ascii_case("subnet6") {
-                            if let Some(uuid) = subnet6.attributes.get("uuid") {
-                                if let Some(subnet_elem) = get_child_ci(subnet6, "subnet") {
-                                    if let Some(cidr) = subnet_elem.get_text() {
-                                        subnets.push(SubnetV6 {
-                                            uuid: uuid.to_string(),
-                                            cidr: cidr.to_string(),
-                                        });
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(subnets)
-}
-
-/// Extract existing Kea reservation IP addresses for duplicate detection
-pub fn extract_existing_reservation_ips(root: &Element) -> Result<HashSet<String>> {
-    let mut ips = HashSet::new();
-
-    // Navigate to <Kea>/<kea> (case-insensitive) -> <dhcp4> -> <reservations>
-    if let Some(kea) = find_descendant_ci(root, "Kea") {
-        if let Some(dhcp4) = find_descendant_ci(kea, "dhcp4") {
-            if let Some(reservations) = find_descendant_ci(dhcp4, "reservations") {
-                for child in reservations.children.iter() {
-                    if let Some(reservation) = child.as_element() {
-                        if reservation.name.eq_ignore_ascii_case("reservation") {
-                            if let Some(ip_elem) = get_child_ci(reservation, "ip_address") {
-                                if let Some(ip) = ip_elem.get_text() {
-                                    ips.insert(ip.to_string());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(ips)
-}
-
-/// Extract existing Kea DHCPv6 reservation IP addresses for duplicate detection
-pub fn extract_existing_reservation_ips_v6(root: &Element) -> Result<HashSet<String>> {
-    let mut ips = HashSet::new();
-
-    if let Some(kea) = find_descendant_ci(root, "Kea") {
-        if let Some(dhcp6) = find_descendant_ci(kea, "dhcp6") {
-            if let Some(reservations) = find_descendant_ci(dhcp6, "reservations") {
-                for child in reservations.children.iter() {
-                    if let Some(reservation) = child.as_element() {
-                        if reservation.name.eq_ignore_ascii_case("reservation") {
-                            if let Some(ip_elem) = get_child_ci(reservation, "ip_address") {
-                                if let Some(ip) = ip_elem.get_text() {
-                                    ips.insert(ip.to_string());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(ips)
-}
-
-/// Extract existing Kea DHCPv6 reservation DUIDs for duplicate detection
-pub fn extract_existing_reservation_duids_v6(root: &Element) -> Result<HashSet<String>> {
-    let mut duids = HashSet::new();
-
-    if let Some(kea) = find_descendant_ci(root, "Kea") {
-        if let Some(dhcp6) = find_descendant_ci(kea, "dhcp6") {
-            if let Some(reservations) = find_descendant_ci(dhcp6, "reservations") {
-                for child in reservations.children.iter() {
-                    if let Some(reservation) = child.as_element() {
-                        if reservation.name.eq_ignore_ascii_case("reservation") {
-                            if let Some(duid_elem) = get_child_ci(reservation, "duid") {
-                                if let Some(duid) = duid_elem.get_text() {
-                                    duids.insert(duid.to_string());
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(duids)
 }
