+++ src/migrate/utils.rs
@@ -1,0 +1,3 @@
+pub(crate) fn short_uuid(uuid: &str) -> &str {
+    uuid.get(..8).unwrap_or(uuid)
+}
